Add a 'hash' mode to pkg repo
Add two new configuration knobs to control creating packages with
the short hash appended to the package name.

PKG_REPO_HASH during repo creation, rename package files into the
"Hashed" directory, with the short hash appended to the filename

PKG_REPO_SYMLINK creates a symlink from the original pkg filename

The hash is separated from the version by a ~ since this character
is not already used as part of a version numbers. Many places in the
code assume the version is everything after the last - character.

Update the fetch code to also use ~ to separate the hash in its cache